fix(ci): fix Linux linker missing Irrlicht transitive deps and pitch clamp direction
- CMakeLists.txt: find_package(JPEG/PNG/ZLIB/BZip2) + find_library(Xxf86vm) on
  Linux; link all to aitown executable. The adrido/irrlicht-vcpkg static library
  does not export these as INTERFACE_LINK_LIBRARIES, causing undefined reference
  to jpeg_*, png_*, BZ2_*, inflate*, and XF86VidMode* at link time.
- CameraController.cpp: fix pitch delta sign (m_pitch -= dy * kRotateSpeed).
  Upward mouse drag (physY decreasing → dy negative) must move pitch toward -20
  (less negative); the previous += produced the opposite direction, causing
  PitchClamp_AtUpperBound and PitchClamp_AtLowerBound tests to swap results.
- main.cpp: cast getAnimators().size() to size_t to silence -Wformat %zu warning.

+++ src/main.cpp
@@ -81,7 +81,7 @@ int main() {
             fprintf(stderr,
                 "[main] WARNING: unexpected animators on addCameraSceneNode() result "
                 "— removing %zu animator(s)\n",
-                cameraNode->getAnimators().size());
+                static_cast<size_t>(cameraNode->getAnimators().size()));
         }
 #endif
         while (cameraNode->getAnimators().size() > 0) {
